longs

+++ Prog/submit/gol.c
@@ -41,10 +41,10 @@ void freeBoard(char ***board, int rows) {
     *board = NULL;
 }
 
-int getAliveCount(struct universe *u) {
-    int count = 0;
-    for(int i = 0;i<u->height;++i) {
-        for(int j = 0;j<u->width;++j){
+long getAliveCount(struct universe *u) {
+    long count = 0;
+    for(long i = 0;i<u->height;++i) {
+        for(long j = 0;j<u->width;++j){
             if(u->board[i][j] == '*'){
                 ++count;
             }
@@ -160,10 +160,10 @@ int will_be_alive(struct universe *u, int column, int row){
 
 int will_be_alive_torus(struct universe *u,  int column, int row){
     int aliveCount = 0;
-    for (int i = row - 1 ; i < row + 2; ++i) {
-        for (int j = column - 1;j < column +2;++j) {
-            int r = i;
-            int c = j;
+    for (long i = row - 1 ; i < row + 2; ++i) {
+        for (long j = column - 1;j < column +2;++j) {
+            long r = i;
+            long c = j;
             if(i == -1) {
                 r = u->height-1;
             }else if(i == u->height) {
